+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Added
 
+- Added  two features to `NavLink` to improve route-based active state without requiring callbacks. #719 by @AnnMarieW
+  - `active="exact-with-search"` matches both pathname and query parameters.
+  - `active="children"` sets parent link to active if any child links are active.
+  
 - Added `buttonProps` to `Button`, `ActionIcon`, `UnstyledButton`, `CopyButton` components to allow passing props directly to the underlying button element. #716 by @AnnMarieW
 
 # 2.6.1